content/xfactor_pipeline/requirements.txt
mediapipe
opencv-python
numpy
Pillow
google-generativeai
playwright
